requirements.txt
aiofiles==23.2.1
aiogram==3.7.0
aiohttp==3.9.5
aiosignal==1.3.1
air_drf_relation==0.6.1
annotated-types==0.7.0
asgiref==3.8.1
attrs==23.2.0
autopep8==2.2.0
certifi==2024.2.2
charset-normalizer==3.3.2
dacite==1.8.1
Django==5.0.6
django-cors-headers==4.3.1
django-environ==0.11.2
django-filter==24.2
django-silk==5.1.0
djangorestframework==3.15.1
djangorestframework-dataclasses==1.3.1
djangorestframework-simplejwt==5.3.1
frozenlist==1.4.1
gprof2dot==2022.7.29
gunicorn==22.0.0
idna==3.7
loguru==0.7.2
magic-filter==1.0.12
multidict==6.0.5
packaging==24.0
psycopg2-binary==2.9.9
pycodestyle==2.11.1
pydantic==2.7.2
pydantic_core==2.18.3
PyJWT==2.8.0
requests==2.32.3
sentry-sdk==2.3.1
sqlparse==0.5.0
typing_extensions==4.12.1
urllib3==2.2.1
yarl==1.9.4
